replace 16 to 32

+++ ui/yas_ui_types.h
@@ -16,7 +16,7 @@ namespace ui {
         vertex2d_t v[4];
     };
 
-    using index2d_t = uint16_t;
+    using index2d_t = uint32_t;
 
     struct index2d_square_t {
         index2d_t v[6];
